+++ app/Keycloak/KeycloakServiceProvider.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Keycloak;
 
 use App\Keycloak\Client\KeycloakHttpClient;
-use App\Keycloak\Service\ApiClient;
+use App\Keycloak\Service\KeycloakApiClient;
 use App\Keycloak\TokenStrategy\ClientCredentials;
 use GuzzleHttp\Client;
 use GuzzleHttp\RequestOptions;
@@ -17,8 +17,8 @@ final class KeycloakServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
-        $this->app->singleton(ApiClient::class, function () {
-            return new ApiClient(
+        $this->app->singleton(KeycloakApiClient::class, function () {
+            return new KeycloakApiClient(
                 new KeycloakHttpClient(
                     new Client([RequestOptions::HTTP_ERRORS => false]),
                     $this->app->get(Config::class),
